rename room -> data_structures
